Created form honeypot challenge
This challenge will use a honeypot technique for form submission.

This is a automated form submission challenge. The form questions will be simple math questions. The user will be required to answer a large number of form questions before they will be provided with the flag.

The Trick: To a normal user using a web browser there will be 5 possible form questions. But in reality there will be 6 form questions, one of which will be hidden from normal user viewing. This is intended to identify sloppy bot design that answers all questions instead of the visible standard ones.